--- awsService/main/main.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"github.com/codeuniversity/smag-mvp/awsService"
-)
-
-func main() {
-	s := awsService.New(9900)
-	s.Listen()
-}
