+++ aws_service/main/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/codeuniversity/smag-mvp/aws_service"
+)
+
+func main() {
+	s := aws_service.New(9900)
+	s.Listen()
+}
